refactor:  1. controller 패키지 추가  2. RacingCarGame class 이동
